separate heardle config
